// Frontend/client/model/documentInformationDto.ts
/**
 * BillingSoftware.Web
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PositionDto } from './positionDto';
import { TypeOfDiscount } from './typeOfDiscount';

export interface DocumentInformationDto { 
    totalDiscount?: number;
    typeOfDiscount?: TypeOfDiscount;
    tax?: number;
    clientId?: string;
    contactPersonId?: string;
    positions?: Array<PositionDto>;
}