// Frontend/client/model/documentInformations.ts
/**
 * BillingSoftware.Web
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Contact } from './contact';
import { Position } from './position';
import { TypeOfDiscount } from './typeOfDiscount';
import { User } from './user';

export interface DocumentInformations { 
    totalDiscount?: number;
    typeOfDiscount?: TypeOfDiscount;
    tax: number;
    clientId?: string;
    contactPersonId?: string;
    readonly totalPriceNet?: number;
    readonly totalPriceGross?: number;
    positions?: Array<Position>;
    contactPerson?: User;
    client?: Contact;
    id?: string;
    rowVersion?: string;
}