// Frontend/client/model/positionDto.ts
/**
 * BillingSoftware.Web
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TypeOfDiscount } from './typeOfDiscount';

export interface PositionDto { 
    quantity?: number;
    discount?: number;
    typeOfDiscount?: TypeOfDiscount;
    productId?: string;
}